Introduce Vue component
