+++ .gitignore
@@ -24,3 +24,4 @@
 /vendor/**/tests/*
 /vendor/**/php4/*
 /vendor/getkirby/composer-installer
+.cache/*
