+++ index.js
@@ -1,243 +1,1 @@
-const PAGE_CREATE_DIALOG = {
-  extends: 'k-page-create-dialog',
-  template: `
-    <k-dialog
-      ref="dialog"
-      :submit-button="$t('page.draft.create')"
-      :notification="notification"
-      size="medium"
-      theme="positive"
-      @submit="$refs.form.submit()"
-    >
-      <k-form
-        ref="form"
-        :fields="fields"
-        :novalidate="true"
-        :key="template"
-        v-model="page"
-        @submit="submit"
-        @input="input"
-      />
-    </k-dialog>
-  `,
-  data() {
-    return {
-      notification: null,
-      parent: null,
-      section: null,
-      templates: [],
-      template: '',
-      page: {},
-      addFields: {}
-    };
-  },
-
-  computed: {
-    fields() {
-      var
-        fields = {},
-        field = {},
-        endpoint = this.$route.path,
-        section = 'addFields';
-
-      if(this.addFields) {
-        fields = this.addFields;
-      } else {
-        fields = {
-          title: {
-            label: this.$t("title"),
-            type: "text",
-            required: true,
-            icon: "title"
-          },
-          slug: {
-            label: this.$t("slug"),
-            type: "text",
-            required: true,
-            counter: false,
-            icon: "url"
-          }
-        }
-      }
-
-      fields.template = {
-        name: "template",
-        label: this.$t("template"),
-        type: "select",
-        disabled: this.templates.length === 1,
-        required: true,
-        icon: "code",
-        empty: false,
-        options: this.templates
-      }
-
-      Object.keys(fields).forEach(name => {
-        field = fields[name];
-
-        if (name != "title" && name != "template" && this.page[name] === undefined){
-          if (field.default !== null && field.default !== undefined) {
-            this.$set(this.page, name, this.$helper.clone(field.default));
-          } else {
-            this.$set(this.page, name, null);
-          }
-        }
-
-        if (name === "title" && this.page[name] === "") {
-          if (field.default !== null && field.default !== undefined) {
-            this.$set(this.page, name, this.$helper.clone(field.default));
-          } else {
-            this.$set(this.page, name, "");
-          }
-        }
-
-        field.section = section;
-        field.endpoints = {
-          field: endpoint + "/addfields/" + this.template + "/" + name,
-          section: endpoint + "/addsections/" + this.template + "/" + section,
-          model: endpoint
-        };
-      });
-
-      return fields;
-    }
-  },
-
-  methods: {
-    open(parent, blueprintApi, section) {
-      this.parent  = parent;
-      this.section = section;
-
-      this.$api
-        .get(blueprintApi + '/add-fields', {section: section})
-        .then(response => {
-          if(response.skipDialog){
-            this.submit(response);
-            return;
-          }
-          this.templates = response.map(blueprint => {
-            return {
-              value: blueprint.name,
-              text: blueprint.title,
-              addFields: blueprint.addFields,
-              options: blueprint.options
-            };
-          });
-
-          if (this.templates[0]) {
-            this.page.template = this.templates[0].value;
-            this.template = this.templates[0].value;
-            this.addFields = this.templates[0].addFields;
-            this.options = this.templates[0].options;
-          }
-
-          this.$refs.dialog.open();
-        })
-        .catch(error => {
-          this.$store.dispatch("notification/error", error);
-        });
-    },
-
-    input() {
-      if(this.page.template !== this.template){
-        var
-          oTemplate = {},
-          template = this.page.template;
-
-        this.template  = template;
-
-        oTemplate = this.templates.find(function(tpl){
-          return tpl.value === template;
-        });
-        this.addFields = oTemplate.addFields;
-        this.options = oTemplate.options;
-        this.$set(this.page, "template", template);
-      }
-    },
-
-    isValid() {
-      var
-        form = this.$refs.form,
-        errors = {},
-        invalid = false;
-
-      if(form) {
-        form.novalidate = false;
-        errors = form.$refs.fields.errors;
-        invalid = true;
-
-        Object.keys(errors).some(field => {
-          var error = errors[field];
-          invalid = error.$pending || error.$invalid || error.$error;
-          return invalid;
-        });
-        return !invalid;
-      } else {
-        return !invalid;
-      }
-    },
-
-    submit(pageData) {
-      if (this.isValid()){
-        var data = {};
-        var route = '';
-
-        if(pageData.skipDialog){
-          data = pageData.page;
-        } else {
-          data = {
-            template: this.page.template,
-            slug: this.page.slug || Date.now(),
-            content: Object.assign({}, this.page)
-          };
-        }
-
-        delete data.content.addFields;
-        delete data.content.template;
-
-        this.$api
-          .post(this.parent + "/children", data)
-          .then(page => {
-            if(this.options && this.options.redirectToNewPage) {
-              route = this.$api.pages.link(page.id);
-            } else {
-              route = page.parent ? this.$api.pages.link(page.parent.id) : '/';
-            }
-
-            this.success({
-              route: route,
-              message: ":)",
-              event: "page.create"
-            });
-            this.$router.go();
-          })
-          .catch(error => {
-            this.$refs.dialog.error(error.message);
-          });
-      } else {
-        this.$refs.dialog.error("Form is not valid");
-      }
-    }
-  }
-};
-
-panel.plugin("steirico/kirby-plugin-custom-add-fields", {
-  components: {
-    'k-page-create-dialog': PAGE_CREATE_DIALOG
-  },
-  use: [
-    function(Vue) {
-      const
-        VUE_COMPONENTS = Vue.options.components;
-
-      Object.keys(VUE_COMPONENTS).forEach(componentName => {
-        const COMPONENT = {
-          components: {
-            'k-page-create-dialog': PAGE_CREATE_DIALOG
-          },
-          extends: VUE_COMPONENTS[componentName]
-        };
-        Vue.component(componentName, COMPONENT);
-      });
-    }
-  ]
-});
+(function () {var b={extends:"k-page-create-dialog",data:function(){return{notification:null,parent:null,section:null,templates:[],template:"",page:{},addFields:{}}},computed:{fields:function(){var e=this,t={},i={},a=this.$route.path;return t=this.addFields?this.addFields:{title:{label:this.$t("title"),type:"text",required:!0,icon:"title"},slug:{label:this.$t("slug"),type:"text",required:!0,counter:!1,icon:"url"}},t.template={name:"template",label:this.$t("template"),type:"select",disabled:1===this.templates.length,required:!0,icon:"code",empty:!1,options:this.templates},Object.keys(t).forEach(function(s){i=t[s],"title"!=s&&"template"!=s&&void 0===e.page[s]&&(null!==i.default&&void 0!==i.default?e.$set(e.page,s,e.$helper.clone(i.default)):e.$set(e.page,s,null)),"title"===s&&""===e.page[s]&&(null!==i.default&&void 0!==i.default?e.$set(e.page,s,e.$helper.clone(i.default)):e.$set(e.page,s,"")),i.section="addFields",i.endpoints={field:a+"/addfields/"+e.template+"/"+s,section:a+"/addsections/"+e.template+"/addFields",model:a}}),t}},methods:{open:function(e,t,i){var a=this;this.parent=e,this.section=i,this.$api.get(t+"/add-fields",{section:i}).then(function(e){e.skipDialog?a.submit(e):(a.templates=e.map(function(e){return{value:e.name,text:e.title,addFields:e.addFields,options:e.options}}),a.templates[0]&&(a.page.template=a.templates[0].value,a.template=a.templates[0].value,a.addFields=a.templates[0].addFields,a.options=a.templates[0].options),a.$refs.dialog.open())}).catch(function(e){a.$store.dispatch("notification/error",e)})},input:function(){if(this.page.template!==this.template){var e,t=this.page.template;this.template=t,e=this.templates.find(function(e){return e.value===t}),this.addFields=e.addFields,this.options=e.options,this.$set(this.page,"template",t)}},isValid:function(){var e=this.$refs.form,t={},i=!1;return e?(e.novalidate=!1,t=e.$refs.fields.errors,i=!0,Object.keys(t).some(function(e){var a=t[e];return i=a.$pending||a.$invalid||a.$error}),!i):!i},submit:function(e){var t=this;if(this.isValid()){var i={},a="";delete(i=e.skipDialog?e.page:{template:this.page.template,slug:this.page.slug||Date.now(),content:Object.assign({},this.page)}).content.addFields,delete i.content.template,this.$api.post(this.parent+"/children",i).then(function(e){a=t.options&&t.options.redirectToNewPage?t.$api.pages.link(e.id):e.parent?t.$api.pages.link(e.parent.id):"/",t.success({route:a,message:":)",event:"page.create"}),t.$router.go()}).catch(function(e){t.$refs.dialog.error(e.message)})}else this.$refs.dialog.error("Form is not valid")}}};if(typeof b==="function"){b=b.options}Object.assign(b,function(){var render=function(){var _vm=this;var _h=_vm.$createElement;var _c=_vm._self._c||_h;return _c("k-dialog",{ref:"dialog",attrs:{"submit-button":_vm.$t("page.draft.create"),"notification":_vm.notification,"size":"medium","theme":"positive"},on:{"submit":function($event){return _vm.$refs.form.submit()}}},[_c("k-form",{key:_vm.template,ref:"form",attrs:{"fields":_vm.fields,"novalidate":true},on:{"submit":_vm.submit,"input":_vm.input},model:{value:_vm.page,callback:function($$v){_vm.page=$$v},expression:"page"}})],1)};var staticRenderFns=[];return{render:render,staticRenderFns:staticRenderFns,_compiled:true,_scopeId:null,functional:undefined}}());panel.plugin("steirico/kirby-plugin-custom-add-fields",{components:{"k-page-create-dialog":b}});})();
